Update distill_modified.py
Edited the distillation process such that we first un-normalize the pixel values w.r.t ImageNet stats (multiply by ImageNet std and add mean), clip to [0, 1], and then re-normalize w.r.t ImageNet stats. This just ensures that the range of pixel values for the synthetic images is the same as real images.